+++ src/App.js
@@ -1,5 +1,6 @@
-import { Navbar, Footer } from "./components";
+import { Navbar, Footer, PetForm } from "./components";
 import { Landing } from "./pages";
+
 import "./App.css";
 import "semantic-ui-css/semantic.min.css";
 import "antd/dist/antd.css";
@@ -14,9 +15,9 @@ function App() {
         <Landing />
         <PetForm />
       </body>
-      <footer>
+      {/* <footer>
         <Footer />
-      </footer>
+      </footer> */}
     </div>
   );
 }
