--- src/components/PetForm.css
@@ -1,5 +1,0 @@
-.form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
