+++ src/components/form/PetForm.css
@@ -1,0 +1,30 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 500px;
+  margin: auto;
+  background-color: #348498;
+  border-radius: 5px;
+  padding: 5px;
+}
+
+.form__header {
+  padding: 5px;
+  display: flex;
+  width: 100%;
+  flex-direction: column;
+  /* align-items: center;
+  justify-content: center; */
+}
+
+.form__dropdown {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
+
+.autocomplete-dropdown-container {
+  width: 250px;
+}
