+++ src/components/index.js
@@ -1,5 +1,6 @@
-import PostCard from './PostCard/PostCard';
-import Navbar from './navbar/Navbar';
-import Footer from './footer/Footer';
+import PostCard from "./PostCard/PostCard";
+import Navbar from "./navbar/Navbar";
+import Footer from "./footer/Footer";
+import PetForm from "./form/PetForm";
 
-export { Navbar, Footer, PostCard };
+export { Navbar, Footer, PostCard, PetForm };
